Import proper .git when extracting a submodule
The new installer submodule does not work on a nixpkgs tree which is
present as a submodule out of the box because the .git directory of a
submodule is located in the superproject.  We work around this by
creating a derivation which copies .git into the nixpkgs tree created
from the submodule.